perf(optimizer): retire dead rules (FPT, MV), keep ConstantFoldHeuristic, label FAT speculative
Optimizer rule hygiene pass after #812 made the rewrite walk total
across all 26 chplan nodes. Each rule re-measured against the full
test/spec corpus (PromQL + LogQL + TraceQL) via an instrumented
fire-probe; verdicts follow the "every technique must earn its place"
standard.

ConstantFoldHeuristic — KEPT. Now fires (1 fixture:
traceql/drilldown_breakdown_kind_rate, whose lowering emits
`(... AND true) AND true` above a MetricsAggregate), reachable only
since #812's total walk reached MetricsAggregate. Earns its place.

MVSubstitution — REMOVED. 0 fires; the default schema ships no live
rollups so the rule was a guaranteed no-op (no rollup roadmap). Deletes
the rule + test + Default()/DefaultWithSchema wiring + 7 synthetic
fixtures + the docs/engine.md MV stage. DefaultWithSchema collapsed into
Default() (its only purpose was binding the rollup registry MV read);
the one caller (test/surface-parity) now uses Default(). The schema-side
Rollup types stay as an unconsumed contract for a future rule, with
their doc comments updated to stop pointing at the deleted file.

FilterProjectTranspose — REMOVED via refactor. 0 fires (no lowering
emits `Filter(Project)`). Its only durable contribution was the shared
`onlyReferencesPassthrough` helper consumed by the two live transpose
rules — hoisted to transpose_shared.go first (go build green), then the
constructor/transform + Default() registration + FPT-only tests deleted.

FilterAggregateTranspose — LABELLED SPECULATIVE. 0 fires; kept as cheap
correctness insurance against a future lowering that surfaces a group-key
filter above an aggregate. doc.go no longer counts it as active.

FilterRangeWindowTranspose — confirmed ACTIVE (1 fire:
promql/subquery_topk_zero_empty, where `topk(0,…)` lowers to
`Filter(false)` above a RangeWindow). The Phase-1 audit's "non-firing"
premise predated #812; the doc reflects the post-total-walk reality
rather than the stale audit.

Plan-neutral: the only test/spec change is the 9 deleted fixtures; no
`-- sql --` / `-- chplan --` / `-- args --` / `-- expected_rows --` cell
in any surviving fixture changed (`just update-golden` default + chdb
both regenerate identically) — proving the removed rules were dead.
