+++ internal/optimizer/decision_pins_test.go
@@ -2,14 +2,11 @@ package optimizer_test
 
 // Cost-based decision pins (Layer 4 § D).
 //
-// The optimizer ships two rules that make non-trivial "should I rewrite"
-// decisions based on plan / schema context:
-//
-//   - MVSubstitution evaluates four safety conditions (step ≥ window,
-//     range ≥ window, range%window == 0, outer-fn commutes with rollup
-//     AggOp). The default v1 cost model is `firstApplicable`.
-//   - The four transpose rules check passthrough columns to decide
-//     whether the rewrite is semantics-preserving.
+// The optimizer's named rules that make non-trivial "should I rewrite"
+// decisions based on plan / schema context are the transpose family
+// (FilterAggregateTranspose, FilterRangeWindowTranspose) and
+// ProjectionPushdown: each checks passthrough columns / column sets to
+// decide whether the rewrite is semantics-preserving.
 //
 // PREWHERE promotion and late-materialisation are not wired as
 // distinct named rules. Their semantic deputies in v1 are the
@@ -25,220 +22,7 @@ import (
 
 	"github.com/tsouza/cerberus/internal/chplan"
 	"github.com/tsouza/cerberus/internal/optimizer"
-	"github.com/tsouza/cerberus/internal/schema"
 )
-
-// mvRule5m is a convenience constructor for the canonical
-// MVSubstitution test rule wiring (5-minute sum-rollup against
-// otel_metrics_sum).
-func mvRule5m() optimizer.Rule {
-	return optimizer.MVSubstitution([]schema.Rollup{sumRollupForInteraction}, "Value")
-}
-
-// mvRule1hAnd5m: registry with the 1h rollup first (coarsest) +
-// 5m fallback. Tests the firstApplicable cost-model picking.
-func mvRule1hAnd5m() optimizer.Rule {
-	return optimizer.MVSubstitution([]schema.Rollup{
-		{
-			BaseTable:   "otel_metrics_sum",
-			RollupTable: "otel_metrics_sum_1h",
-			Window:      time.Hour,
-			AggOp:       schema.RollupAggSum,
-			ValueColumn: "Sum",
-		},
-		sumRollupForInteraction,
-	}, "Value")
-}
-
-// rangeWindow is a typed helper that builds a RangeWindow with the
-// supplied func + range + step over otel_metrics_sum.
-func rangeWindow(fn string, rng, step time.Duration) *chplan.RangeWindow {
-	return &chplan.RangeWindow{
-		Input:           &chplan.Scan{Table: "otel_metrics_sum"},
-		Func:            fn,
-		Range:           rng,
-		Step:            step,
-		TimestampColumn: "TimeUnix",
-		ValueColumn:     "Value",
-		GroupBy:         []chplan.Expr{&chplan.ColumnRef{Name: "Attributes"}},
-	}
-}
-
-// scanTableFromRangeWindow extracts the rewritten Scan's Table from
-// the rule output, fatalling if the shape is wrong.
-func scanTableFromRangeWindow(t *testing.T, n chplan.Node) string {
-	t.Helper()
-	rw, ok := n.(*chplan.RangeWindow)
-	if !ok {
-		t.Fatalf("expected RangeWindow at root, got %T", n)
-	}
-	scan, ok := rw.Input.(*chplan.Scan)
-	if !ok {
-		t.Fatalf("expected Scan child, got %T", rw.Input)
-	}
-	return scan.Table
-}
-
-// TestDecision_MVSub_StepEqualsWindowApplies pins the boundary case
-// (1): step == window is the minimum step the rollup buckets allow.
-// Must apply.
-func TestDecision_MVSub_StepEqualsWindowApplies(t *testing.T) {
-	t.Parallel()
-	plan := rangeWindow("sum_over_time", time.Hour, 5*time.Minute)
-	out, _ := mvRule5m().Apply(plan)
-	if got := scanTableFromRangeWindow(t, out); got != "otel_metrics_sum_5m" {
-		t.Fatalf("step=5m == window=5m must apply; got Scan.Table=%q", got)
-	}
-}
-
-// TestDecision_MVSub_RangeMultipleApplies pins safety condition (2):
-// range = 2 × window. Must apply.
-func TestDecision_MVSub_RangeMultipleApplies(t *testing.T) {
-	t.Parallel()
-	plan := rangeWindow("sum_over_time", 10*time.Minute, 5*time.Minute)
-	out, _ := mvRule5m().Apply(plan)
-	if got := scanTableFromRangeWindow(t, out); got != "otel_metrics_sum_5m" {
-		t.Fatalf("range=10m = 2×window=5m must apply; got Scan.Table=%q", got)
-	}
-}
-
-// TestDecision_MVSub_RangeOffByOneRejected covers the "almost
-// aligned" case: range = window + 1 unit. Mod is non-zero so must
-// reject.
-func TestDecision_MVSub_RangeOffByOneRejected(t *testing.T) {
-	t.Parallel()
-	plan := rangeWindow("sum_over_time", 5*time.Minute+time.Second, 5*time.Minute)
-	out, _ := mvRule5m().Apply(plan)
-	if got := scanTableFromRangeWindow(t, out); got != "otel_metrics_sum" {
-		t.Fatalf("range=5m1s not multiple of 5m must reject; got Scan.Table=%q", got)
-	}
-}
-
-// TestDecision_MVSub_DifferentAggregatorRejected covers safety
-// condition (3): query uses `count_over_time` but the only rollup
-// is a sum-rollup. Must reject — sum ≠ count.
-func TestDecision_MVSub_DifferentAggregatorRejected(t *testing.T) {
-	t.Parallel()
-	plan := rangeWindow("count_over_time", time.Hour, 5*time.Minute)
-	out, _ := mvRule5m().Apply(plan)
-	if got := scanTableFromRangeWindow(t, out); got != "otel_metrics_sum" {
-		t.Fatalf("count_over_time over sum-rollup must reject; got Scan.Table=%q", got)
-	}
-}
-
-// TestDecision_MVSub_RateRejected pins the v1 rate exclusion:
-// rate(metric[5m]) over a sum-rollup is NOT safe (rate looks at
-// per-sample deltas which the sum-rollup loses). Must reject.
-func TestDecision_MVSub_RateRejected(t *testing.T) {
-	t.Parallel()
-	plan := rangeWindow("rate", time.Hour, 5*time.Minute)
-	out, _ := mvRule5m().Apply(plan)
-	if got := scanTableFromRangeWindow(t, out); got != "otel_metrics_sum" {
-		t.Fatalf("rate over sum-rollup must reject (per-sample deltas not preserved); got Scan.Table=%q", got)
-	}
-}
-
-// TestDecision_MVSub_CoarsestFirst pins the firstApplicable cost
-// model: with both 1h and 5m rollups available and the query
-// permitting both, pick 1h.
-func TestDecision_MVSub_CoarsestFirst(t *testing.T) {
-	t.Parallel()
-	plan := rangeWindow("sum_over_time", 24*time.Hour, time.Hour)
-	out, _ := mvRule1hAnd5m().Apply(plan)
-	if got := scanTableFromRangeWindow(t, out); got != "otel_metrics_sum_1h" {
-		t.Fatalf("firstApplicable must pick coarsest (1h); got Scan.Table=%q", got)
-	}
-}
-
-// TestDecision_MVSub_FallbackWhenCoarsestRejected: 1h is rejected
-// (step < window), 5m must be picked.
-func TestDecision_MVSub_FallbackWhenCoarsestRejected(t *testing.T) {
-	t.Parallel()
-	plan := rangeWindow("sum_over_time", time.Hour, 5*time.Minute)
-	out, _ := mvRule1hAnd5m().Apply(plan)
-	if got := scanTableFromRangeWindow(t, out); got != "otel_metrics_sum_5m" {
-		t.Fatalf("expected 1h rejected (step<window), fall through to 5m; got Scan.Table=%q", got)
-	}
-}
-
-// TestDecision_TransposeFilter_PushedUnderProject_BareColumn pins
-// the canonical "predicate references a passthrough column" → rewrite.
-// This is the transpose equivalent of "PREWHERE on the sort-prefix
-// promotes" — the predicate sits next to the Scan after the rewrite.
-func TestDecision_TransposeFilter_PushedUnderProject_BareColumn(t *testing.T) {
-	t.Parallel()
-	scan := &chplan.Scan{Table: "otel_metrics_gauge"}
-	plan := &chplan.Filter{
-		Input: &chplan.Project{
-			Input: scan,
-			Projections: []chplan.Projection{
-				{Expr: &chplan.ColumnRef{Name: "MetricName"}},
-			},
-		},
-		Predicate: labelFilter("MetricName", "up"),
-	}
-	out, _ := optimizer.FilterProjectTranspose().Apply(plan)
-	proj, ok := out.(*chplan.Project)
-	if !ok {
-		t.Fatalf("expected Project at root, got %T", out)
-	}
-	if _, ok := proj.Input.(*chplan.Filter); !ok {
-		t.Fatalf("expected Filter under Project (predicate pushed below the projection layer), got %T", proj.Input)
-	}
-}
-
-// TestDecision_TransposeFilter_NotPushed_ComputedColumn pins the
-// inverse decision: predicate references a computed Project column
-// → don't push. The Filter stays above the Project.
-func TestDecision_TransposeFilter_NotPushed_ComputedColumn(t *testing.T) {
-	t.Parallel()
-	scan := &chplan.Scan{Table: "otel_metrics_gauge"}
-	plan := &chplan.Filter{
-		Input: &chplan.Project{
-			Input: scan,
-			Projections: []chplan.Projection{
-				{
-					Expr: &chplan.Binary{
-						Op:    chplan.OpMul,
-						Left:  &chplan.ColumnRef{Name: "Value"},
-						Right: &chplan.LitInt{V: 2},
-					},
-					Alias: "doubled",
-				},
-				{Expr: &chplan.ColumnRef{Name: "MetricName"}},
-			},
-		},
-		// Predicate references `doubled` — only exists post-Project.
-		Predicate: &chplan.Binary{
-			Op:    chplan.OpGt,
-			Left:  &chplan.ColumnRef{Name: "doubled"},
-			Right: &chplan.LitInt{V: 10},
-		},
-	}
-	out, _ := optimizer.FilterProjectTranspose().Apply(plan)
-	if _, ok := out.(*chplan.Filter); !ok {
-		t.Fatalf("expected Filter at root (rewrite must be declined for computed column); got %T", out)
-	}
-}
-
-// TestDecision_TransposeFilter_NotPushed_RenamedColumn: predicate
-// touches an aliased name that has no source-side equivalent.
-func TestDecision_TransposeFilter_NotPushed_RenamedColumn(t *testing.T) {
-	t.Parallel()
-	plan := &chplan.Filter{
-		Input: &chplan.Project{
-			Input: &chplan.Scan{Table: "otel_metrics_gauge"},
-			Projections: []chplan.Projection{
-				{Expr: &chplan.ColumnRef{Name: "MetricName"}, Alias: "metric"},
-			},
-		},
-		Predicate: labelFilter("metric", "up"),
-	}
-	out, _ := optimizer.FilterProjectTranspose().Apply(plan)
-	if _, ok := out.(*chplan.Filter); !ok {
-		t.Fatalf("expected Filter at root (rewrite must be declined for renamed column); got %T", out)
-	}
-}
 
 // TestDecision_RangeWindowTranspose_PushedOnSeriesKey: predicate
 // references a bare series-key column (Attributes) → pushed under
