--- test/spec/optimizer/filter_project_transpose_blocked.txtar
@@ -1,12 +1,0 @@
--- optimized --
-SELECT * FROM (SELECT `MetricName` AS `metric` FROM (SELECT `MetricName` FROM `otel_metrics_gauge`)) WHERE (`metric` = ?)
--- unoptimized --
-SELECT * FROM (SELECT `MetricName` AS `metric` FROM (SELECT * FROM `otel_metrics_gauge`)) WHERE (`metric` = ?)
--- chplan_optimized --
-Filter predicate=(metric = "up")
-  Project [MetricName AS metric]
-    Scan(otel_metrics_gauge, columns=[MetricName])
--- chplan_unoptimized --
-Filter predicate=(metric = "up")
-  Project [MetricName AS metric]
-    Scan(otel_metrics_gauge)
