--- test/spec/optimizer/filter_project_transpose_passes.txtar
@@ -1,12 +1,0 @@
--- optimized --
-SELECT `MetricName`, `Value` FROM (SELECT `MetricName`, `Value` FROM `otel_metrics_gauge` WHERE (`MetricName` = ?))
--- unoptimized --
-SELECT * FROM (SELECT `MetricName`, `Value` FROM (SELECT * FROM `otel_metrics_gauge`)) WHERE (`MetricName` = ?)
--- chplan_optimized --
-Project [MetricName, Value]
-  Filter predicate=(MetricName = "up")
-    Scan(otel_metrics_gauge, columns=[MetricName, Value])
--- chplan_unoptimized --
-Filter predicate=(MetricName = "up")
-  Project [MetricName, Value]
-    Scan(otel_metrics_gauge)
