--- test/spec/optimizer/mv_sub_through_safe_filter.txtar
@@ -1,12 +1,0 @@
--- unoptimized --
-SELECT * FROM (SELECT `Attributes`, arraySum(window_vals) AS `Value` FROM (SELECT `Attributes`, arrayMap(p -> tupleElement(p, 2), window_pairs) AS `window_vals`, arraySum(arrayMap((p, c) -> if(c < p, c, c - p), arrayPopBack(arrayMap(x -> tupleElement(x, 2), window_pairs)), arrayPopFront(arrayMap(x -> tupleElement(x, 2), window_pairs)))) AS `counter_delta` FROM (SELECT `Attributes`, arrayFilter(p -> tupleElement(p, 1) > now64(9) - toIntervalNanosecond(3600000000000) AND tupleElement(p, 1) <= now64(9), series_array) AS `window_pairs` FROM (SELECT `Attributes`, arraySort(groupArray((`TimeUnix`, `Value`))) AS `series_array` FROM (SELECT * FROM `otel_metrics_sum`) GROUP BY `Attributes`))) WHERE length(`window_vals`) >= 1) WHERE (`Attributes`[?] = ?)
--- optimized --
-SELECT `Attributes`, arraySum(window_vals) AS `Sum` FROM (SELECT `Attributes`, arrayMap(p -> tupleElement(p, 2), window_pairs) AS `window_vals`, arraySum(arrayMap((p, c) -> if(c < p, c, c - p), arrayPopBack(arrayMap(x -> tupleElement(x, 2), window_pairs)), arrayPopFront(arrayMap(x -> tupleElement(x, 2), window_pairs)))) AS `counter_delta` FROM (SELECT `Attributes`, arrayFilter(p -> tupleElement(p, 1) > now64(9) - toIntervalNanosecond(3600000000000) AND tupleElement(p, 1) <= now64(9), series_array) AS `window_pairs` FROM (SELECT `Attributes`, arraySort(groupArray((`TimeUnix`, `Sum`))) AS `series_array` FROM (SELECT * FROM `otel_metrics_sum_5m` WHERE (`Attributes`[?] = ?)) GROUP BY `Attributes`))) WHERE length(`window_vals`) >= 1
--- chplan_unoptimized --
-Filter predicate=(Attributes["job"] = "api")
-  RangeWindow func=sum_over_time range=1h0m0s step=5m0s ts=TimeUnix value=Value groupBy=[Attributes]
-    Scan(otel_metrics_sum)
--- chplan_optimized --
-RangeWindow func=sum_over_time range=1h0m0s step=5m0s ts=TimeUnix value=Sum groupBy=[Attributes]
-  Filter predicate=(Attributes["job"] = "api")
-    Scan(otel_metrics_sum_5m)
