--- test/spec/optimizer/mv_substitution_sum_5m.txtar
@@ -1,10 +1,0 @@
--- unoptimized --
-SELECT `Attributes`, arraySum(window_vals) AS `Value` FROM (SELECT `Attributes`, arrayMap(p -> tupleElement(p, 2), window_pairs) AS `window_vals`, arraySum(arrayMap((p, c) -> if(c < p, c, c - p), arrayPopBack(arrayMap(x -> tupleElement(x, 2), window_pairs)), arrayPopFront(arrayMap(x -> tupleElement(x, 2), window_pairs)))) AS `counter_delta` FROM (SELECT `Attributes`, arrayFilter(p -> tupleElement(p, 1) > now64(9) - toIntervalNanosecond(3600000000000) AND tupleElement(p, 1) <= now64(9), series_array) AS `window_pairs` FROM (SELECT `Attributes`, arraySort(groupArray((`TimeUnix`, `Value`))) AS `series_array` FROM (SELECT * FROM `otel_metrics_sum`) GROUP BY `Attributes`))) WHERE length(`window_vals`) >= 1
--- optimized --
-SELECT `Attributes`, arraySum(window_vals) AS `Sum` FROM (SELECT `Attributes`, arrayMap(p -> tupleElement(p, 2), window_pairs) AS `window_vals`, arraySum(arrayMap((p, c) -> if(c < p, c, c - p), arrayPopBack(arrayMap(x -> tupleElement(x, 2), window_pairs)), arrayPopFront(arrayMap(x -> tupleElement(x, 2), window_pairs)))) AS `counter_delta` FROM (SELECT `Attributes`, arrayFilter(p -> tupleElement(p, 1) > now64(9) - toIntervalNanosecond(3600000000000) AND tupleElement(p, 1) <= now64(9), series_array) AS `window_pairs` FROM (SELECT `Attributes`, arraySort(groupArray((`TimeUnix`, `Sum`))) AS `series_array` FROM (SELECT * FROM `otel_metrics_sum_5m`) GROUP BY `Attributes`))) WHERE length(`window_vals`) >= 1
--- chplan_optimized --
-RangeWindow func=sum_over_time range=1h0m0s step=5m0s ts=TimeUnix value=Sum groupBy=[Attributes]
-  Scan(otel_metrics_sum_5m)
--- chplan_unoptimized --
-RangeWindow func=sum_over_time range=1h0m0s step=5m0s ts=TimeUnix value=Value groupBy=[Attributes]
-  Scan(otel_metrics_sum)
